Step3

+++ src/main/java/com/jongmin/mystorage/service/file/FileService.java
@@ -1,6 +1,5 @@
 package com.jongmin.mystorage.service.file;
 
-import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.core.io.Resource;
@@ -8,11 +7,9 @@ import org.springframework.stereotype.Service;
 
 import com.jongmin.mystorage.controller.api.dto.UploadFileRequestDto;
 import com.jongmin.mystorage.exception.FileAlreadyExistException;
-import com.jongmin.mystorage.exception.FileNotInDatabaseException;
 import com.jongmin.mystorage.exception.FileNotInFileSystemException;
 import com.jongmin.mystorage.model.MyFile;
 import com.jongmin.mystorage.model.MyFolder;
-import com.jongmin.mystorage.model.enums.FileItemStatus;
 import com.jongmin.mystorage.repository.FileRepository;
 import com.jongmin.mystorage.service.response.FileResponse;
 import com.jongmin.mystorage.service.response.StringResponse;
@@ -34,29 +31,10 @@ public class FileService {
 	private final FileRepositoryUtils fileRepositoryUtils;
 	private final FileIoUtils fileIoUtils;
 
-	public MyFile checkMyFileAndGet(String ownerName, UUID fileUuid) {
-		Optional<MyFile> optional = fileRepository.findByUuid(fileUuid);
-		if (optional.isEmpty()) {
-			throw new FileNotInDatabaseException("파일에 대한 정보가 DB에 존재하지 않습니다.");
-		}
-		MyFile myFile = optional.get();
-		if (myFile.getStatus() != FileItemStatus.SAVED) {
-			throw new RuntimeException("파일이 삭제되어 있는 상태입니다.");
-		}
-		if (!myFile.getOwnerName().equals(ownerName)) {
-			throw new RuntimeException("본인 소유의 파일이 아닙니다.");
-		}
-		if (fileIoUtils.fileNotExists(myFile)) {
-			throw new FileNotInFileSystemException("파일이 디스크 상에 존재하지 않습니다.");
-		}
-
-		return myFile;
-	}
-
 	public FileResponse uploadFile(String ownerName, UploadFileRequestDto requestDto) {
 		MyFolder parentFolder = folderRepositoryUtils.getFolderByUuid(requestDto.getFolderUuid());
 		MyFile myFileEntity = MyFile.createMyFileEntity(requestDto.getMultipartFile(), ownerName, parentFolder);
-		if (fileRepositoryUtils.fileExists(myFileEntity)) {
+		if (fileRepositoryUtils.sameFileNameExistsInFolder(myFileEntity, parentFolder)) {
 			throw new FileAlreadyExistException("이미 동일한 이름의 파일이 존재합니다.");
 		}
 
@@ -66,7 +44,7 @@ public class FileService {
 	}
 
 	public FileResponse readFile(String ownerName, UUID fileUuid) {
-		MyFile checkedFile = checkMyFileAndGet(ownerName, fileUuid);
+		MyFile checkedFile = fileRepositoryUtils.getFileByUuidWithSavedStatus(ownerName, fileUuid);
 		return FileResponse.fromMyFile(checkedFile);
 	}
 
@@ -74,17 +52,38 @@ public class FileService {
 		// 파일을 삭제를 위한 Lock을 얻고,
 		// 파일 상태를 확인하고 삭제하는 과정이 atomic하게 진행될 수 있도록 추후에 수정이 필요합니다.
 		// 파일을 다운로드 받는 사용자가 있는 경우 역시 고려되어야 합니다.
-		MyFile checkedFile = checkMyFileAndGet(ownerName, fileUuid);
-		fileIoUtils.deleteFile(checkedFile);
-		fileRepositoryUtils.deleteFile(checkedFile);
+		MyFile checkedFile = fileRepositoryUtils.getFileByUuidWithSavedStatus(ownerName, fileUuid);
 
-		return new StringResponse("요청한 파일에 대한 삭제가성공적으로 진행되었습니다.");
+		if (fileIoUtils.fileNotExists(checkedFile)) {
+			throw new FileNotInFileSystemException("파일이 디스크 상에 존재하지 않습니다.");
+		}
+		fileRepositoryUtils.deleteFile(checkedFile);
+		fileIoUtils.deleteFile(checkedFile);
+
+		return new StringResponse("요청한 파일에 대한 삭제가 성공적으로 진행되었습니다.");
 	}
 
 	public Resource downloadFile(String ownerName, UUID fileUuid) {
 		// 파일을 다운로드를 위한 Lock을 얻고,
 		// 파일 상태를 확인하고 다운로드 받는 과정이 atomic하게 진행될 수 있도록 추후에 수정이 필요합니다.
-		MyFile checkedFile = checkMyFileAndGet(ownerName, fileUuid);
+		MyFile checkedFile = fileRepositoryUtils.getFileByUuidWithSavedStatus(ownerName, fileUuid);
+		if (fileIoUtils.fileNotExists(checkedFile)) {
+			throw new FileNotInFileSystemException("파일이 디스크 상에 존재하지 않습니다.");
+		}
 		return fileIoUtils.fileToResource(checkedFile);
+	}
+
+	@Transactional
+	public FileResponse moveFile(String ownerName, UUID fileUuid, UUID destFolderUuid) {
+		MyFile file = fileRepositoryUtils.getFileByUuidWithSavedStatus(ownerName, fileUuid);
+		MyFolder destFolder = folderRepositoryUtils.getFolderByUuidWithSavedStatus(ownerName, destFolderUuid);
+
+		if (fileRepositoryUtils.sameFileNameExistsInFolder(file, destFolder)) {
+			throw new FileAlreadyExistException("옮기려는 폴더에 동일한 이름의 파일이 존재해 이동이 불가능 합니다.");
+		} else {
+			file.move(destFolder);
+		}
+
+		return FileResponse.fromMyFile(file);
 	}
 }
